Fix the network type selection to 2G if another SIM is on >2G
This is a hackish way to handle the device hardware limitation
(only one SIM card connected to a >2G network at once), the user
should be informed of the limitation and still be offered to
register their preference.

This is an adaptation of the same change for Android 6 (that fixed
FPIIM-300) with a simpler execution.

Issue: FP2N-107
Change-Id: I942ee1d51ad3b5cec7076341d7b1d4ec86a20cfe