Restrict selection of preferred network type based on second SIM state
With a single SIM card inserted in the phone, selecting 2G
as network mode type results in inability to switch again to
a different network mode without a network settings reset.

This commit restricts the possibility to choose a network mode
different from 2G only if the second SIM in a non-active state
or if the second SIM is in 2G mode itself.

Issue: FP2N-313
Change-Id: I5aa05c7c2a2c56eede68fa8e318b530dc2d4c308