Drop mobile data page from setup wizard [3/5]
This was a workaround for earlier Android version and wasn't actually
intended to be ported to Android 11. Besides, it's integrated at a point
in the setup wizard and the mobile data setup flow where it's not really
supported by AOSP/GMS and causing mobile data connectivity issues in
some scenarios.

Revert "Add new Activity in SetupWizard for Mobile data settings"

This reverts commit df203b81ab9b673e50048ca32b6968e09e72cc9b.

Issue: FP3-A11#357
Change-Id: Ie21ec477444e3ab637cd8db512c10a9286436900